fix: move function to private method
